move partial components to /components directory
