+++ components/layouts/Episode.tsx
@@ -1,9 +1,8 @@
 import Head from "next/head";
 import EmbedAnchor from "../EmbedAnchor";
 import styles from "../../styles/episode.module.scss";
-import Services from "../components/partial/Footer";
-import Header from "../partial/Header";
-import Footer from "../partial/Footer";
+import Header from "../Header";
+import Footer from "../Footer";
 
 export default function Episode({ children, frontMatter }) {
   const epNum = frontMatter.__resourcePath.match(/^ep\/([0-9]+)-/)[1];
